Moved the functions for accessing the database via sqlite3 to the data folder and refactored the relevant associations
