+++ src/data/manager_service.py
@@ -1,11 +1,6 @@
 import sqlite3
 import dotenv
-
-
-def get_db_connection():
-    conn = sqlite3.connect(f'../../data/inhouse.db')
-    conn.execute("PRAGMA foreign_keys = ON")
-    return conn
+import db_access
 
 
 def set_env_variable(key, newvalue):
@@ -21,34 +16,34 @@ def delete_env_variable(key):
 
 
 def get_reference_list(identifier=None):
-    conn = get_db_connection()
+    conn = db_access.get_db_connection()
     if not identifier:
         login_list = list(conn.cursor().execute("""SELECT Title, ServerId FROM SteamLogin""").fetchall())
     else:
         login_list = list(conn.cursor().execute("""SELECT Title, ServerId FROM SteamLogin
                                         WHERE Title = ? OR ServerId = ?""", [identifier, identifier]).fetchall())
-    conn.close()
+    db_access.close_db_connection(conn)
     return login_list
 
 
 def get_reference_server(identifier):
-    conn = get_db_connection()
+    conn = db_access.get_db_connection()
     conn.row_factory = lambda cursor, row: row[0]
     login_list = list(conn.cursor().execute("""SELECT ServerId FROM SteamLogin
                                             WHERE Title = ? OR ServerId = ?""", [identifier, identifier]).fetchall())
-    conn.close()
     server = str(login_list[0])
+    db_access.close_db_connection(conn)
     return server
 
 
 def delete_credentials(identifier):
-    conn = get_db_connection()
+    conn = db_access.get_db_connection()
     server = get_reference_server(identifier)
     delete_env_variable(server + "username")
     delete_env_variable(server + "password")
     conn.cursor().execute("""DELETE FROM SteamLogin WHERE Title = ? OR ServerId = ?""", [identifier, identifier])
     conn.commit()
-    conn.close()
+    db_access.close_db_connection(conn)
     return
 
 
@@ -60,28 +55,28 @@ def edit_credentials(identifier, username, password):
 
 
 def add_credentials(title, server, username, password):
-    conn = get_db_connection()
+    conn = db_access.get_db_connection()
     try:
         conn.cursor().execute("""INSERT INTO SteamLogin (Title, ServerId) Values (?, ?)""",
                               [title, server])
         set_env_variable(server + "username", username)
         set_env_variable(server + "password", password)
         conn.commit()
-        conn.close()
+        db_access.close_db_connection(conn)
     except sqlite3.IntegrityError:
         print("Server does not exist within the database!")
     return
 
 
 def database_exists():
-    conn = get_db_connection()
+    conn = db_access.get_db_connection()
     check = conn.cursor().execute("""SELECT name FROM sqlite_master; """).fetchall()
-    conn.close()
+    db_access.close_db_connection(conn)
     return check
 
 
 def create_tables():
-    conn = get_db_connection()
+    conn = db_access.get_db_connection()
     conn.cursor().execute("""CREATE TABLE IF NOT EXISTS User(Id INTEGER PRIMARY KEY, Discord BIGINT UNIQUE, Steam BIGINT UNIQUE,
         MMR INT, Pos1 INT, Pos2 INT,Pos3 INT, Pos4 INT, Pos5 INT, LastUpdated DATETIME)""")
     conn.cursor().execute("""CREATE TABLE IF NOT EXISTS Server(Id INTEGER PRIMARY KEY, Server BIGINT, AdminChannel BIGINT,
@@ -98,4 +93,4 @@ def create_tables():
         LeagueId INT, ViewerDelay INT, FOREIGN KEY(ServerId) REFERENCES Server(Id))""")
     conn.cursor().execute("""CREATE TABLE IF NOT EXISTS SteamLogin(ServerId INT UNIQUE, Title CHAR,
         FOREIGN KEY(ServerId) REFERENCES Server(Id))""")
-    conn.close()
+    db_access.close_db_connection(conn)
